+++ testing/splash.js
@@ -10,7 +10,7 @@
 * Github: https://github.com/cameramanben/LUTCalc
 */
 lutInputs.addInput('version','v2.60');
-lutInputs.addInput('date','August 2016');
+lutInputs.addInput('date','September 2016');
 var splash = document.getElementById('javascriptwarning');
 splash.removeChild(splash.firstChild);
 splash.className = 'splash';